+++ docs/src/Image.doc.js
@@ -1,6 +1,11 @@
 // @flow
 import * as React from 'react';
 import { Image } from 'gestalt';
+import stock1 from './images/stock1.jpg';
+import stock2 from './images/stock2.jpg';
+import stock3 from './images/stock3.jpg';
+import stock4 from './images/stock4.jpg';
+import stock5 from './images/stock5.jpg';
 import PropTable from './components/PropTable';
 import Example from './components/Example';
 import PageHeader from './components/PageHeader';
@@ -118,7 +123,7 @@ card(
   color="rgb(111, 91, 77)"
   naturalHeight={564}
   naturalWidth={564}
-  src="https://i.pinimg.com/474x/b2/55/ed/b255edbf773ffb3985394e6efb9d2a49.jpg"
+  src="${stock1}"
 />
 </Column>
 `}
@@ -139,7 +144,7 @@ card(
   color="rgb(231, 186, 176)"
   naturalHeight={751}
   naturalWidth={564}
-  src="https://i.pinimg.com/564x/19/f4/87/19f487a680f9fb1ecc8aa139b2afac7f.jpg"
+  src="${stock2}"
 >
   <Box padding={3}>
     <Text color="white">
@@ -197,7 +202,7 @@ card(
         fit="contain"
         naturalHeight={1}
         naturalWidth={1}
-        src="https://i.pinimg.com/564x/a9/dd/08/a9dd080b383ba4f336b5e4705cacdfba.jpg"
+        src="${stock3}"
       />
     </Box>
     <Box
@@ -213,7 +218,7 @@ card(
         fit="cover"
         naturalHeight={1}
         naturalWidth={1}
-        src="https://i.pinimg.com/564x/a9/dd/08/a9dd080b383ba4f336b5e4705cacdfba.jpg"
+        src="${stock3}"
       />
     </Box>
   </Box>
@@ -234,7 +239,7 @@ card(
         fit="contain"
         naturalHeight={1}
         naturalWidth={1}
-        src="https://i.pinimg.com/564x/47/b8/ec/47b8ec3948ef1f8eba0b0fe1dde28622.jpg"
+        src="${stock4}"
       />
     </Box>
     <Box
@@ -250,7 +255,7 @@ card(
         fit="cover"
         naturalHeight={1}
         naturalWidth={1}
-        src="https://i.pinimg.com/564x/47/b8/ec/47b8ec3948ef1f8eba0b0fe1dde28622.jpg"
+        src="${stock4}"
       />
     </Box>
   </Box>
@@ -271,7 +276,7 @@ card(
         fit="contain"
         naturalHeight={1}
         naturalWidth={1}
-        src="https://i.pinimg.com/564x/10/7d/99/107d993c7818ed41bc0afc27ab2a2f15.jpg"
+        src="${stock5}"
       />
     </Box>
     <Box
@@ -287,7 +292,7 @@ card(
         fit="cover"
         naturalHeight={1}
         naturalWidth={1}
-        src="https://i.pinimg.com/564x/10/7d/99/107d993c7818ed41bc0afc27ab2a2f15.jpg"
+        src="${stock5}"
       />
     </Box>
   </Box>