+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 
 ### Patch
 
+* Internal: update `yarn.lock` (#152)
+* Docs: include images in repo (#151)
+
 </details>
 
 ## 0.68.0 (May 3, 2018)
@@ -17,7 +20,6 @@
 ### Minor
 
 * Button / SearchField / SegmentedControl / SelectList / Tabs / TextField: consistent sizing + improve Windows compatibility (#148)
-
 * Icon: Add new prop to Icon -- dangerouslySetSvgPath (#142)
 
 ## 0.67.0 (April 25, 2018)